:recycle: [REFACTOR] pop 기존코드 활성화
