+++ BackEnd/favshare/src/main/java/com/favshare/idol/dto/IdolDto.java
@@ -1,4 +1,4 @@
-package com.favshare._temp.dto;
+package com.favshare.idol.dto;
 
 import com.favshare.idol.entity.IdolEntity;
