+++ BackEnd/favshare/src/main/java/com/favshare/pop/dto/pop/PopAlgoDto.java
@@ -1,4 +1,4 @@
-package com.favshare._temp.dto;
+package com.favshare.pop.dto.pop;
 
 import lombok.*;
