+++ BackEnd/favshare/src/main/java/com/favshare/pop/dto/pop/PopInfoResponse.java
@@ -1,7 +1,5 @@
 package com.favshare.pop.dto.pop;
 
-import com.favshare._temp.dto.PopInfoDto;
-import com.favshare._temp.dto.input.UserProfileDto;
 import lombok.*;
 
 @Getter
