+++ BackEnd/favshare/src/main/java/com/favshare/pop/dto/pop/UserProfileDto.java
@@ -1,4 +1,4 @@
-package com.favshare._temp.dto.input;
+package com.favshare.pop.dto.pop;
 
 import com.favshare.user.entity.User;
