+++ BackEnd/favshare/src/main/java/com/favshare/user/service/UserService.java
@@ -10,15 +10,15 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import com.favshare._temp.dto.FeedDto;
-import com.favshare._temp.dto.IdolDto;
-import com.favshare._temp.dto.PopDto;
+import com.favshare.idol.dto.IdolDto;
+import com.favshare.pop.dto.pop.PopDto;
 import com.favshare._temp.dto.UserAccountDto;
 import com.favshare._temp.dto.input.EmailPasswordDto;
 import com.favshare._temp.dto.input.FeedUserIdDto;
 import com.favshare._temp.dto.input.FollowForFollowDto;
-import com.favshare._temp.dto.input.FriendFeedDto;
+import com.favshare.pop.dto.pop.FriendFeedDto;
 import com.favshare._temp.dto.input.UserInfoDto;
-import com.favshare._temp.dto.input.UserProfileDto;
+import com.favshare.pop.dto.pop.UserProfileDto;
 import com.favshare._temp.dto.input.UserSignUpDto;
 import com.favshare.follow.entity.FollowEntity;
 import com.favshare.idol.entity.IdolEntity;
