+++ BackEnd/favshare/src/main/java/com/favshare/youtube/dto/YoutubeDetail.java
@@ -1,6 +1,6 @@
 package com.favshare.youtube.dto;
 
-import com.favshare._temp.dto.PopDto;
+import com.favshare.pop.dto.pop.PopDto;
 import lombok.Data;
 import lombok.NoArgsConstructor;
